Prevent deletion of default banner and profile images if user decides to update them

+++ Posts/posts_/controllers/delete_post.py
@@ -15,7 +15,8 @@ def delete_post(post_id: str):
 
     post_to_delete: PostModel = get_post(post_id)
 
-    os.remove(UPLOAD_FOLDER_PATH, post_to_delete.post_file)
+    if post_to_delete.post_file is not None:
+        os.remove(UPLOAD_FOLDER_PATH, post_to_delete.post_file)
 
     if post_to_delete.parent is not None:
         post_to_delete.parent.comments -= 1
